- Правильно организовал код для лучшей читаемости.
